gpt oss without ep
